Bump to 5001, part 1

+++ src/cpuexec.h
@@ -43,7 +43,7 @@ struct MachineCPU
 {
 	int			cpu_type;					/* index for the CPU type */
 	int			cpu_flags;					/* flags; see #defines below */
-	int			cpu_clock;					/* in Hertz */
+	double		cpu_clock;					/* in Hertz */
 	const void *memory_read;				/* struct Memory_ReadAddress */
 	const void *memory_write;				/* struct Memory_WriteAddress */
 	const void *port_read;
@@ -51,7 +51,7 @@ struct MachineCPU
 	void 		(*vblank_interrupt)(void);	/* for interrupts tied to VBLANK */
 	int 		vblank_interrupts_per_frame;/* usually 1 */
 	void 		(*timed_interrupt)(void);	/* for interrupts not tied to VBLANK */
-	int 		timed_interrupts_per_second;
+	double 		timed_interrupts_per_second;
 	void *		reset_param;				/* parameter for cpu_reset */
 	const char *tag;
 };
